Fixed a reference error of toast wrapper
I have met a bad case about lifecycle when `Toast` component working with `Redirect` component of the lib `react-router`. 
With the switch of the route, the toastWrapper of `theme` has gone, but the lifecycle still not stopped. So I suggest to check this wrapper if it is exists to prevent throwing errors.