fix: Do not toggle NavigationView with boolean
If a boolean is used for the toggleExpanded function and the variable is the same as the current state (i.e. telling the NavigationView to close when it is already closed, or open when it is already open), the NavigationView would have toggled it's state in the past.
This (hopefully) fixes this issue.